% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ign_urba_assembler_sf.R
\name{ign_urba_assembler_sf}
\alias{ign_urba_assembler_sf}
\title{Assembler les objets de classe sf contenus dans une liste en un unique objet sf.}
\usage{
ign_urba_assembler_sf(liste_sf)
}
\arguments{
\item{liste_sf}{Une liste d'objets de classe sf.}
}
\value{
Un objet unique de classe sf qui est la fusion des objets de la liste
}
\description{
Assembler les objets de classe sf contenus dans une liste en un unique objet sf.
}
\examples{
\dontrun{
assemblage <- ign_urba_assembler_sf(liste_sf = sf_liste)
}
}
