% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ign_urba_dec.R
\name{ign_urba_dec}
\alias{ign_urba_dec}
\title{Décompresser les fichiers zip de l'API portail de l'urbanisme.}
\usage{
ign_urba_dec(repertoire)
}
\arguments{
\item{repertoire}{Texte. Chemin vers le répertoire où sont stockés les fichiers
compressés.}
}
\value{
Autant de sous-répertoire qu'il y a de fichiers .zip.
}
\description{
Décompresser les fichiers zip de l'API portail de l'urbanisme.
}
\examples{
\dontrun{
dec_ign_urba(repertoire = "raw_data/prescription_surf")
}
}
