% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ign_urba_lire_shapes.R
\name{ign_urba_lire_shapes}
\alias{ign_urba_lire_shapes}
\title{Lire les fichiers shapefiles contenus dans un répertoire}
\usage{
ign_urba_lire_shapes(repertoire)
}
\arguments{
\item{repertoire}{Texte. Chemin vers le répertoire où sont stockés les shapefiles.}
}
\value{
Une liste contenant un objet de classe sf par shapefile.
}
\description{
Lire les fichiers shapefiles contenus dans un répertoire
}
\examples{
\dontrun{
sf_liste <- ign_urba_lire_shapes(repertoire = "raw_data/prescription_surf")
}
}
