% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ign_urba_net_rep.R
\name{ign_urba_net_rep}
\alias{ign_urba_net_rep}
\title{Nettoyer les répertoires de données brutes (suppression des shapefiles vides).}
\usage{
ign_urba_net_rep(repertoire_donnees_brutes, seuil_ko = 2)
}
\arguments{
\item{repertoire_donnees_brutes}{Texte. Chemin vers le répertoire où sont stockés les shapefiles.}

\item{seuil_ko}{Numérique. Valeur seuil de taille des fichiers shapefiles au-dessous
de laquelle ils seront considérés comme vides et supprimés.}
}
\value{
Le répertoire de données expurgé des éventuels shapefiles vides.
}
\description{
Nettoyer les répertoires de données brutes (suppression des shapefiles vides).
}
\examples{
\dontrun{
ign_urba_net_rep(repertoire_donnees_brutes = "raw_data/prescription_surf")
}
}
