% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ign_urba_sauver_shape.R
\name{ign_urba_sauver_shape}
\alias{ign_urba_sauver_shape}
\title{Exporter un objet de classe sf en shapefile.}
\usage{
ign_urba_sauver_shape(
  objet_sf,
  scr = NA,
  repertoire_sortie,
  nom_fichier_sortie
)
}
\arguments{
\item{objet_sf}{L'objet sf à exporter.}

\item{scr}{Numérique. Numéro EPSG du système de coordonnées. La valeur par défaut
est 4326, ce qui correspond au WGS84.}

\item{repertoire_sortie}{Texte. Chemin vers le fichier répertoire où le fichier
sera écrit.}

\item{nom_fichier_sortie}{Texte. Nom du fichier de sortie avec son extension .shp.}
}
\value{
L'objet géographique au format shapefile.
}
\description{
Exporter un objet de classe sf en shapefile.
}
\examples{
\dontrun{
# Sauvegarde en Lambert 93 (code EPSG 2154)
ign_urba_sauver_shape(objet_sf = assemblage,
                      scr = 2154,
                      repertoire_sortie = "processed_data"
                      nom_fichier_sortie = "prescription_lin.shp")
}
}
