% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/menv_bh_sta_tod.R
\name{menv_bh_sta_tod}
\alias{menv_bh_sta_tod}
\title{Collecter les informations de base sur les stations de mesure.}
\usage{
menv_bh_sta_tod(
  bbox = NULL,
  code_commune_site = NULL,
  code_cours_eau = NULL,
  code_departement = NULL,
  code_region = NULL,
  code_site = NULL,
  code_troncon_hydro_site = NULL,
  code_zone_hydro_site = NULL,
  distance = NULL,
  fields = NULL,
  latitude = NULL,
  libelle_cours_eau = NULL,
  libelle_site = NULL,
  longitude = NULL,
  page = NULL,
  size = 1000
)
}
\arguments{
\item{bbox}{Voir lien ci-dessus.}

\item{code_commune_site}{Voir lien ci-dessus.}

\item{code_cours_eau}{Idem.}

\item{code_departement}{Idem.}

\item{code_region}{Idem.}

\item{code_site}{Idem.}

\item{code_troncon_hydro_site}{Idem.}

\item{code_zone_hydro_site}{Idem.}

\item{distance}{Idem.}

\item{fields}{Idem.}

\item{latitude}{Idem.}

\item{libelle_cours_eau}{Idem.}

\item{libelle_site}{Idem.}

\item{longitude}{Idem.}

\item{page}{Idem.}

\item{size}{Idem.}
}
\value{
Un dataframe avec les caractéristiques des stations sélectionnées.
}
\description{
Tous les paramètres sont décrits sur la page de l'API :
    https://hubeau.eaufrance.fr/page/api-hydrometrie#/hydrometrie/sites
}
\examples{
\dontrun{
donnees_stations <- menv_bh_sta_tod(code_region = 53)
}
}
