% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ofb_bnpe_pp_tod.R
\name{ofb_bnpe_pp_tod}
\alias{ofb_bnpe_pp_tod}
\title{Collecter les informations sur les points de prélèvement de la BNPE.}
\usage{
ofb_bnpe_pp_tod(
  bbox = NULL,
  code_bss_point_eau = NULL,
  code_commune_insee = NULL,
  code_departement = NULL,
  code_nature = NULL,
  code_ouvrage = NULL,
  code_point_prelevement = NULL,
  code_type_milieu = NULL,
  code_zone_hydro = NULL,
  date_exploitation = NULL,
  fields = NULL,
  format = "json",
  libelle_departement = NULL,
  nappe_accompagnement = NULL,
  nom_commune = NULL,
  nom_point_prelevement = NULL,
  page = 1,
  size = 1000
)
}
\arguments{
\item{bbox}{Voir lien ci-dessus.}

\item{code_bss_point_eau}{Idem.}

\item{code_commune_insee}{Idem.}

\item{code_departement}{Idem.}

\item{code_nature}{Idem.}

\item{code_ouvrage}{Idem.}

\item{code_point_prelevement}{Idem.}

\item{code_type_milieu}{Idem.}

\item{code_zone_hydro}{Idem.}

\item{date_exploitation}{Idem.}

\item{fields}{Idem.}

\item{format}{Idem.}

\item{libelle_departement}{Idem.}

\item{nappe_accompagnement}{Idem.}

\item{nom_commune}{Idem.}

\item{nom_point_prelevement}{Idem.}

\item{page}{Idem.}

\item{size}{Idem.}
}
\value{
Un dataframe avec les caractéristiques des points de prélèvement sélectionnés.
}
\description{
Tous les paramètres sont décrits sur la page de l'API :
    https://hubeau.eaufrance.fr/page/api-prelevements-eau#console
}
\examples{
\dontrun{
donnees_stations <- ofb_bnpe_pp_tod(code_departement = 29)
}
}
