% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osm_creer_polygone_region.R
\name{osm_creer_polygone_region}
\alias{osm_creer_polygone_region}
\title{Créer un polygone pour une région}
\usage{
osm_creer_polygone_region(
  departements_sf,
  departements_selectionnes,
  distance_buffer = 0,
  intitule_region
)
}
\arguments{
\item{departements_sf}{Objet sf des contours des départements. Typiquement obtenu depuis la
fonction osm_depts_tod().}

\item{departements_selectionnes}{Caractère. Vecteur des numéros des départements.}

\item{distance_buffer}{Numérique. Paramètre qui détermine la largeur du buffer en km.
(approximativement).}

\item{intitule_region}{Caractère. Intitulé de la région.}
}
\value{
Objet de classe sf décrivant le contour de la région y compris le buffer.
}
\description{
Créer un polygone de classe sf à partir des numéros des départements et
    de l'objet sf des départements
}
\examples{
\dontrun{
region <- osm_creer_polygone_region(departements_sf = depts,
    departements_selectionnes = c("22", "29", "35", "56"),
    distance_buffer = 0.01,
    intitule_region = "Bretagne")
}
}
