% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/req_api_geo_urba.R
\name{req_api_geo_urba}
\alias{req_api_geo_urba}
\title{Télécharger des couches shapefile depuis le géoportail de l'urbanisme.}
\usage{
req_api_geo_urba(
  couche,
  index_debut = 0,
  nb_elements_par_telech = 10000,
  n_tot_elements_a_telech = 1e+06,
  repertoire = NA
)
}
\arguments{
\item{couche}{Texte. Nom de la couche à télécharger.}

\item{index_debut}{Numérique. Index du premier des éléments géographiques à télécharger.}

\item{nb_elements_par_telech}{Numérique. Nombre d'éléments à télécharger par paquet.}

\item{n_tot_elements_a_telech}{Numérique. Si l'on le connaît, le nombre total d'éléments
géographiques à télécharger.}

\item{repertoire}{Texte. Chemin vers le répertoire de stockage des fichiers téléchargés.
Par défaut, c'est un sous répertoire de "raw_data" nommé d'après la couche. S'il n'existe
pas, la fonction le crée.}
}
\value{
Les fichiers téléchargés.
}
\description{
Si l'emprise géographique est vaste, le nombre d'éléments géographiques excède la
    limite de l'API (10 000), auquel cas le téléchargement est effectué par "paquets"
    dont la taille est paramétrable.
}
\details{
ATTENTION : d'éventuels anciens fichiers portant les mêmes noms que ceux téléchargés
    dans le même répertoire seraient écrasés.
}
\examples{
\dontrun{
req_api_geo_urba(couche = "wfs_du:prescription_lin",
repertoire = "raw_data/req_api_geo_urba")
}
}
