% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sandre_referentiel_taxo_tod.R
\name{sandre_referentiel_taxo_tod}
\alias{sandre_referentiel_taxo_tod}
\title{Importer une table de correspondance entre les codes espèces 3 lettres Aspe et leurs codes Taxref}
\usage{
sandre_referentiel_taxo_tod(
  url = "https://api.sandre.eaufrance.fr/referentiels/v1/apt.csv?compress=true"
)
}
\arguments{
\item{url}{Caractère. URL du jeu de données à télécharger.
Par défaut "https://api.sandre.eaufrance.fr/referentiels/v1/apt.csv?compress=true"}
}
\value{
Dataframe. Table de correspondance
}
\description{
La donnée ne provient pas de la base ASPE, mais de l'API SANDRE. Comme elle importe l'ensemble du
    référentiel taxonomique, l'opération prend plusieurs dizaines de secondes.
}
\examples{
\dontrun{
taxref <- sandre_referentiel_taxo_tod()
}
}
