% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_geobretagne.R
\name{sf_geobretagne}
\alias{sf_geobretagne}
\title{Requêter les flux WFS proposés sur https://geobretagne.fr}
\usage{
sf_geobretagne(
  url_wfs = "https://geobretagne.fr/geoserver/dreal_b/sage_dreal/wfs?",
  couche,
  repertoire_donnees_brutes = NA
)
}
\arguments{
\item{url_wfs}{Caractère. URL du flux WFS.}

\item{couche}{Caractère. Nom de la couche à charger.}

\item{repertoire_donnees_brutes}{Texte. Chemin vers le répertoire de stockage des fichiers téléchargés.
S'il n'est pas renseigné, la donnée brute n'est pas sauvegardée.}
}
\value{
Un objet de classe sf correspondant à la couche choisie.
}
\description{
Requêter les flux WFS proposés sur https://geobretagne.fr
}
\examples{
\dontrun{
bassins <- tgeo_wfs_sandre(url_wfs = "https://services.sandre.eaufrance.fr/geo/sandre?",
                           couche = "BassinDCE")
}
}
