% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sie_carthage_bassins_agr.R
\name{sie_carthage_bassins_agr}
\alias{sie_carthage_bassins_agr}
\title{Aggréger les entités hydrographiques}
\usage{
sie_carthage_bassins_agr(
  bassin_geo_poly,
  echelle = c("Zone_Hydro", "Sous_Secteur_Hydro", "Secteur_Hydro", "Region_Hydro",
    "District_hydro"),
  prop_pts_a_garder = 1
)
}
\arguments{
\item{bassin_geo_poly}{Objet de classe sf contenant les polygones délimitant les
Zones HYdro. Typiquement produit par la fonction tod::sie_carthage_bassins_tod().}

\item{echelle}{Caractère. Au choix entre "Zone_Hydro", "Sous_Secteur_Hydro", "Secteur_Hydro"
et "Region_Hydro".}

\item{prop_pts_a_garder}{Nombre entre zéro et 1 (défaut 1) indiquant la proportions de points
à conserver pour simplifier les contours une fois les Zones Hydro agrégées.}
}
\value{
Objet de classe sf contenant les polygones au niveau de découpage des
    entités hydrographiques choisi, simplifié sauf si prop_pts_a_garder = 1.
}
\description{
A partir du découpage le plus fin proposé par la fonction sie_carthage_bassins_tod()
    du package {tod}, qui est volumineux (+ de 6000 polygones), permet d'agréger les
    Zones HYdro en Sous-secteurs Hydro, Secteurs Hydro ou Régions hydro.
}
\details{
NB : Pour l'échelle du district hydrographique, il est nécessaire d'avoir attribué à chaque
    polygone le code et le libellé de son district dans des variables respectivement
    nommées "CdDistrictHy" et "LbDistrictHy".
}
\examples{
\dontrun{
regions_hydro <- sie_carthage_bassins_agr(
bassin_geo_poly = zones_hydro,
echelle = "Region_Hydro",
pc_pts_a_garder = 0.01)
}
}
