% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sie_carthage_bassins_tod.R
\name{sie_carthage_bassins_tod}
\alias{sie_carthage_bassins_tod}
\title{Créer un objet sf avec polygones du découpage hydrographique à partir de geo.data.gouv.fr}
\usage{
sie_carthage_bassins_tod(url = NA, repertoire = "raw_data")
}
\arguments{
\item{url}{Texte. URL de téléchargement de la couche shapefile.}

\item{repertoire}{Texte. Chemin vers le répertoire de stockage et de décompression.}
}
\value{
Un objet sf multi polygones.
}
\description{
Créer un objet sf avec polygones du découpage hydrographique à partir de geo.data.gouv.fr
}
\examples{
\dontrun{
bv <- sie_carthage_bassins_tod(repertoire = "raw_data")
}
}
