% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tgeo_ign_urba.R
\name{tgeo_ign_urba}
\alias{tgeo_ign_urba}
\title{Créer une couches shapefile depuis le géoportail de l'urbanisme.}
\usage{
tgeo_ign_urba(
  couche,
  ymin,
  xmin,
  ymax,
  xmax,
  scr = 4326,
  index_debut = 0,
  nb_elements_par_telech = 10000,
  n_tot_elements_a_telech = 1e+06,
  repertoire_donnees_brutes = "raw_data",
  repertoire_sortie = NA,
  nom_fichier_sortie = "ign_urba.shp",
  seuil_ko = 2
)
}
\arguments{
\item{couche}{Texte. Nom de la couche à télécharger.}

\item{ymin}{Numérique. Latitude minimale de l'emprise géographique en WGS84 (degrés, point décimal).}

\item{xmin}{Numérique. Longitude minimale de l'emprise géographique en WGS84 (degrés, point décimal).}

\item{ymax}{Numérique. Latitude maximale de l'emprise géographique en WGS84 (degrés, point décimal).}

\item{xmax}{Numérique. Longitude maximale de l'emprise géographique en WGS84 (degrés, point décimal).}

\item{scr}{Numérique. Numéro EPSG du système de coordonnées. La valeur par défaut
est 4326, ce qui correspond au WGS84.}

\item{index_debut}{Numérique. Index du premier des éléments géographiques à télécharger.}

\item{nb_elements_par_telech}{Numérique. Nombre d'éléments à télécharger par paquet.}

\item{n_tot_elements_a_telech}{Numérique. Si l'on le connaît, le nombre total d'éléments
géographiques à télécharger.}

\item{repertoire_donnees_brutes}{Texte. Chemin vers le répertoire de stockage des fichiers téléchargés.
Par défaut, c'est un sous-répertoire de "raw_data" nommé d'après la couche. S'il n'existe
pas, la fonction le crée.}

\item{repertoire_sortie}{Texte. Chemin vers le fichier répertoire où le fichier
sera écrit.}

\item{nom_fichier_sortie}{Texte. Nom du fichier de sortie avec son extension .shp.}

\item{seuil_ko}{Numérique. Valeur seuil de taille des fichiers shapefiles au-dessous
de laquelle ils seront considérés comme vides et supprimés.}
}
\value{
Les fichiers téléchargés.
}
\description{
Si l'emprise géographique est vaste, le nombre d'éléments géographiques excède la
    limite de l'API (10 000), auquel cas le téléchargement est effectué par "paquets"
    dont la taille est paramétrable.
}
\details{
ATTENTION : d'éventuels anciens fichiers portant les mêmes noms que ceux téléchargés
    dans le même répertoire seraient écrasés.
}
\examples{
\dontrun{
ma_couche <- "prescription_lin"
repertoire_donnees_brutes <- "raw_data/prescription_lin"
fichier_donnees_traitees <- "processed_data/prescription_lin.shp"
tgeo_ign_urba(couche = ma_couche,
                 xmin = -7,
                 ymin = 47,
                 xmax = -4,
                 ymax = 48,
                 repertoire_donnees_brutes = repertoire_donnees_brutes,
                 fichier_sortie = fichier_donnees_traitees)
}
}
