% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tgeo_osm_adm.R
\name{tgeo_osm_adm}
\alias{tgeo_osm_adm}
\title{Obtenir le découpage administratif de la France}
\usage{
tgeo_osm_adm(
  url,
  repertoire_donnees_brutes = NA,
  nom_fichier_zip = "download.zip"
)
}
\arguments{
\item{url}{Texte. URL de téléchargement.}

\item{repertoire_donnees_brutes}{Texte. Chemin vers le répertoire où le fichier sera stocké.}

\item{nom_fichier_zip}{Texte. Nom du fichier zip à sauvegarder.
Par défaut c'est "download.zip".}
}
\value{
Un objet sf polygone des contours
}
\description{
Cette fonction télécharge depuis la plateforme data.gouv.fr les contours du
    découpage administratif (source : OpenStreetMap) qui est ensuite décompressé
    et converti en objet de classe sf. Si le répertoire de stockage des données brutes
    n'est pas indiqué, elles ne sont pas sauvegardées.
}
\examples{
\dontrun{
url <- "https://www.data.gouv.fr/fr/datasets/r/3096e551-c68d-40ce-8972-a228c94c0ad1"
depts <- tgeo_osm_adm(url = url, repertoire_donnees_brutes = "raw_data/depts")
}
}
