% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ign_urba_tod.R
\name{tod_ign_urba}
\alias{tod_ign_urba}
\title{Télécharger des couches shapefile depuis le géoportail de l'urbanisme.}
\usage{
tod_ign_urba(
  couche,
  ymin,
  xmin,
  ymax,
  xmax,
  index_debut = 0,
  nb_elements_par_telech = 10000,
  n_tot_elements_a_telech = 1e+06,
  repertoire = NA
)
}
\arguments{
\item{couche}{Texte. Nom de la couche à télécharger.}

\item{ymin}{Numérique. Latitude minimale de l'emprise géographique en WGS84 (degrés, point décimal).}

\item{xmin}{Numérique. Longitude minimale de l'emprise géographique en WGS84 (degrés, point décimal).}

\item{ymax}{Numérique. Latitude maximale de l'emprise géographique en WGS84 (degrés, point décimal).}

\item{xmax}{Numérique. Longitude maximale de l'emprise géographique en WGS84 (degrés, point décimal).}

\item{index_debut}{Numérique. Index du premier des éléments géographiques à télécharger.}

\item{nb_elements_par_telech}{Numérique. Nombre d'éléments à télécharger par paquet.}

\item{n_tot_elements_a_telech}{Numérique. Si l'on le connaît, le nombre total d'éléments
géographiques à télécharger.}

\item{repertoire}{Texte. Chemin vers le répertoire de stockage des fichiers téléchargés.
Par défaut, c'est un sous répertoire de "raw_data" nommé d'après la couche. S'il n'existe
pas, la fonction le crée.}
}
\value{
Les fichiers téléchargés.
}
\description{
Si l'emprise géographique est vaste, le nombre d'éléments géographiques excède la
    limite de l'API (10 000), auquel cas le téléchargement est effectué par "paquets"
    dont la taille est paramétrable.
}
\details{
ATTENTION : d'éventuels anciens fichiers portant les mêmes noms que ceux téléchargés
    dans le même répertoire seraient écrasés.
}
\examples{
\dontrun{
tod_ign_urba(couche = "prescription_lin",
xmin = -7,
ymin = 47,
xmax = -2,
ymax = 48,
repertoire = "raw_data/tod_ign_urba")
}
}
