% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttab_dept_reg.R
\name{ttab_dept_reg}
\alias{ttab_dept_reg}
\title{Table de correspondance départements - régions}
\usage{
ttab_dept_reg(
  url = "https://geo.api.gouv.fr/departements?fields=nom,code,codeRegion,region",
  repertoire_donnees_brutes = NA
)
}
\arguments{
\item{url}{Caractère. URL de l'API géo "départements"}
}
\value{
Un dataframe indiquant l'appartenance des départements aux régions
}
\description{
Téléchargée depuis l'API Géo
}
\examples{
\dontrun{
correspondance <- ttab_dept_reg()
depts <- depts \%>\%
  left_join(correspondance, by = c("code_insee" = "code"))
}
}
