% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ugeo_aggr.R
\name{ugeo_aggr}
\alias{ugeo_aggr}
\title{Agréger des polygones sf selon un facteur}
\usage{
ugeo_aggr(polygones_sf, variable_aggregation)
}
\arguments{
\item{polygones_sf}{Objet sf contenant les objets à agréger}

\item{variable_aggregation}{Caractère. Nom de la variable qui sert à constituer les
agrégats (par exemple le code région).}
}
\value{
Objet sf agrégé.
}
\description{
Agréger des polygones sf selon un facteur
}
\examples{
\dontrun{
bretagne <- depts \%>\%
  filter(code \%in\% c("22", "29", "35", "56")) \%>\%
  ugeo_fusion(distance_buffer = 0.01)
}
}
