% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ugeo_fusion.R
\name{ugeo_fusion}
\alias{ugeo_fusion}
\title{Fusionner des polygones sf}
\usage{
ugeo_fusion(polygones_sf, distance_buffer)
}
\arguments{
\item{polygones_sf}{Objet sf contenant les objets à fusionner.}

\item{distance_buffer}{Distance du tampon à constituer autour de l'objet fusionné
(en km, approximatif).}
}
\value{
Le polygone issu de la fusion.
}
\description{
Fusionner des polygones sf
}
\examples{
\dontrun{
bretagne <- depts \%>\%
  filter(code \%in\% c("22", "29", "35", "56")) \%>\%
  ugeo_fusion(distance_buffer = 0.01)
}
}
