% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wfs_geobretagne.R
\name{wfs_geobretagne}
\alias{wfs_geobretagne}
\title{Accéder au flux WFS GéoBretagne}
\usage{
wfs_geobretagne(
  url_wfs = "https://geobretagne.fr/geoserver/dreal_b/",
  couche = "sage_dreal"
)
}
\arguments{
\item{url_wfs}{Texte. URL du flux.}

\item{couche}{Texte. Nom de la couche. Par défaut "sage_dreal".}
}
\value{
L'objet de classe sf retourné par le flux. Par défaut les SAGEs de Bretagne.
}
\description{
Accéder au flux WFS GéoBretagne
}
\examples{
\dontrun{
sages <- tod::sf_geobretagne()
mapview::mapview(sages)
}
}
