% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_dept_reg.R
\name{df_dept_reg}
\alias{df_dept_reg}
\title{Correspondance départements - régions}
\usage{
df_dept_reg(
  url = "https://geo.api.gouv.fr/departements?fields=nom,code,codeRegion,region"
)
}
\arguments{
\item{url}{Caractère. URL de l'API géo "départements"}
}
\value{
Un dataframe indiquant l'appartenance des départements aux régions
}
\description{
Correspondance départements - régions
}
\examples{
\dontrun{
correspondance <- df_dept_reg()
depts <- depts \%>\%
  left_join(correspondance, by = c("code_insee" = "code"))
}
}
