% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/menv_bh_htr_tod.R
\name{menv_bh_htr_tod}
\alias{menv_bh_htr_tod}
\title{Télécharger les données hydrologiques en temps réel sur une station. Plus d'informations sur
    https://hubeau.eaufrance.fr/page/api-hydrometrie#/hydrometrie/observations}
\usage{
menv_bh_htr_tod(station_id, mesure = "Q", debut = NA, fin = NA)
}
\arguments{
\item{station_id}{Caractères. Identifiant de la station comme "J7010610".}

\item{mesure}{Texte. Grandeur mesurée qui peut prendre les valeurs "Q" pour le débit ou
"H" pour les hauteurs d'eau.}

\item{debut}{Caractère. Date de début de la série au format "jj/mm/aaaa". Par défaut le
paramètre fait remonter sur une semaine à partir de l'argument 'fin'.}

\item{fin}{Caractère. Date de fin de la série au format "jj/mm/aaaa". Par défaut c'est la
date du jour selon l'horloge interne de la machine.}
}
\value{
Dataframe.
}
\description{
Le nom de la fonction correspond à "menv" pour la source, le ministère de l'environnement,
    "bh" pour la banque HYDRO, "htr" pour "hydrologie temps réel" et "tod" pour
    "téléchargement open data".
}
\examples{
\dontrun{
data_sta <- menv_bh_htr_tod(station_id = "J7010610")
ggplot(data = data_sta, aes(x = date_obs, y = resultat_obs)) +
geom_point()
}
}
