% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/menv_bh_htr_tod.R
\name{menv_bh_htr_tod}
\alias{menv_bh_htr_tod}
\title{Extension de la fonction menv_bh_htr_tod()}
\usage{
menv_bh_htr_tod(stations_id, grandeur_hydro = "Q", timestep = 60)
}
\arguments{
\item{stations_id}{Vecteur texte. Identifiants des stations.}

\item{grandeur_hydro}{Texte. Grandeur hydrométrique observée pour la série. Prend la valeur
"Q" pour le débit et "H" pour la hauteur d'eau.}

\item{timestep}{Pas de temps entre mesures consécutives.}
}
\value{
Le dataframe contenant les données.
}
\description{
Permet de télécharger plus de données à la fois, mais admet moins d'arguments.
}
\examples{
\dontrun{
mes_stations <- c("J7010610", "J4125710")
debits_plusieurs_stations <- menv_bh_htr_tod(stations_id = mes_stations)
}
}
