% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osm_depts_tod.R
\name{osm_depts_tod}
\alias{osm_depts_tod}
\title{Télécharger les contours des départements}
\usage{
osm_depts_tod(url, repertoire)
}
\arguments{
\item{url}{URL de téléchargement.}

\item{repertoire}{Chemin vers le répertoire où le fichier sera stocké.}
}
\value{
Le shapefile des contours de départements.
}
\description{
Cette fonction télécharge depuis la plateforme data.gouv.fr une couche simplifiée
    des contours des départements et la décompresse. La source est OpenStreetMap.
}
\examples{
\dontrun{
url <- "https://www.data.gouv.fr/fr/datasets/r/3096e551-c68d-40ce-8972-a228c94c0ad1"
depts <- osm_depts_tod(url = url, repertoire = "raw_data")
}
}
